requirements.txt
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.4
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
